Refactor main - use more functions
Split some of the functionality out of main() in to separate functions, in preparation
for adding unit tests at some point.

Currently just dropping the documents table and recreating it on each run. Would be nice
to handle this better, e.g. with a separate setup and with code to handle duplicates.
But this will do as an MVP for now.