Switch question from Maru to the Indian election
Since the election happened after llama3 was trained, this better demonstrates that
the model is using RAG.

Also provided a general wiki link about the article, where the results are towards the
bottom of the page.

When tested by running the code, the output was good, accurately stating that Modi won
with a 240-seat majority.